+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-core
 
+## 0.1.6
+
 ## 0.1.4
 
 ### Patch Changes